### 2019-12-12 #1954 @whedon
[PRE REVIEW]: LFSpy: A Python Implementation of Local Feature Selection for Data Classification with scikit-learn Compatibility
**Submitting author:** @kiretd (<a href="http://orcid.org/0000-0003-4849-732X">Kiret Dhindsa</a>)
**Repository:** <a href="https://github.com/McMasterRS/LFSpy" target ="_blank">https://github.com/McMasterRS/LFSpy</a>
**Version:** 1.0.0
