### 2019-12-12 #1954 @kyleniemeyer
Hi @kiretd, could you review the references and add the missing DOIs? (Including other references that whedon may have missed)
