### 2019-12-12 #1954 @dfm
@kiretd: as well as fixing the DOIs, please take a look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and let me know if you see anyone who would be an especially good fit (don't @ them!). I'll start pinging reviewers in the next few days one way or another. Thanks!
