+++ paper.bib
@@ -6,7 +6,8 @@
   number={6},
   pages={1217--1227},
   year={2015},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPAMI.2015.2478471}
 }
 
 @article{Armanfard:2017,
@@ -17,7 +18,8 @@
   number={5},
   pages={1396--1413},
   year={2017},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TNNLS.2017.2676101}
 }
 
 @article{scikit-learn,
@@ -37,5 +39,6 @@
   author={Armanfard, Narges and Komeili, Majid and Reilly, James P and Connoly, John},
   journal={IEEE journal of biomedical and health informatics},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/JBHI.2018.2877738}
 }
