### 2019-12-13 #1954 @kiretd
Hi @dfm and @kyleniemeyer, thanks for such a quick turnaround!
I've added the DOIs for all references.
Going through the list of reviewers, here are some candidates that look like a really good fit to me:
desilinguist, rougier, sealhuang, krother, ahurriyetoglu, nirum, arokem, effigies, oesteban, jsgalan, tupi, mnarayan, tianxzhu, GalenStocking, ryEllison
