### 2019-12-13 353c74c
Adding test.py for verification on two datasets
