### 2019-12-13 b93000b
added argument for seeding the random rounding
