### 2019-12-15 #1954 @dfm
:wave: Hey @effigies! Any chance you're available and willing to review this JOSS submission? Thanks!
