### 2019-12-16 14abee4
add testing info, will need to update with import update
