### 2019-12-16 4e84729
Refactored tests for compatibility with pytest
