### 2019-12-16 c5e4ada
fix to pytest tests
