### 2019-12-16 8162a30
Added requirements to setup.py
