### 2019-12-16 4103a77
Adjusted the test files to make sure they can run as a part of the library
