### 2019-12-16 9f20eb4
Finished converting tests to pytest format by updating the setup.py file
