+++ LFSpy/LFS.py
@@ -1,0 +1,515 @@
+# LFS: local feature selection and classification algorithm
+#  Citation: [1] and [2]
+# [1] N. Armanfard, JP. Reilly, and M. Komeili, "Local Feature Selection for Data Classification", IEEE Trans. on Pattern Analysis and Machine Intelligence, vol. 38, no. 6, pp. 1217-1227, 2016.
+# [2] N. Armanfard, JP. Reilly, and M. Komeili, "Logistic Localized Modeling of the Sample Space for Feature Selection and Classification", IEEE Transactions on Neural Networks and Learning Systems, vol. 29, no. 5, pp. 1396-1413, 2018
+# --------------------------------------------------------------------------
+# INPUT:
+#        Train (M by N):  training data : [x1,x2,...xN] Each column is an observation; M is number of candidate features.
+#        TrainLables (1 by N): class label = {0,1}.
+#        Test (M by K): M is number of candidate features; K is number of test points.
+#        TestLables (1 by K): class label = {0,1}.
+#        Para:  parameters.
+#            Para.gamma:   impurity level (default: 0.2).
+#            Para.tau:    number of iterations (default: 2).
+#            Para.sigma: controlls neighboring samples weighting (default: 1).
+#            Para.alpha:  maximum number of selected feature for each representative point.
+#            Para.NBeta:  numer of distinct \beta (default: 20).
+#            Para.NRRP:   number of iterations for randomized rounding process (defaul 2000).
+# OUTPUT:
+#        fstar (M by N):  selected features for each representative point; if fstar(i,j)=1, jth feature is selected for ith representative point.
+#        fstarLin (M by N): it is fstar before applying randomized rounding process.
+#        ErCls1: it is the classification error (in percent) associated to the input test points with class label 1.
+#        ErCls2: it is the classification error (in percent) associated to the input test points with class label 0.
+#        ErClassification: it is the total error (in percent) for the entire input test points.
+# --------------------------------------------------------------------------
+#  by Narges Armanfard (pythonized by Oliver Cook)
+#  update history: April 22 2019 (Matlab Code) August 30th 2019 (Python)
+#==========================================================================
+
+import numpy as np
+from scipy.io import loadmat
+from scipy.optimize import linprog
+
+WANDERING = 0
+OBSERVATIONS = 0
+FEATURES = 1
+
+def distance_between(A, B, axis=None):
+    return np.abs(np.sqrt(np.sum((A - B))**2, axis=axis))
+
+###########################################
+# Local Feature Selection
+###########################################
+
+class LocalFeatureSelection():
+
+    def __init__(self, alpha=19, gamma=0.2, tau=2, sigma=1, n_beta=20, nrrp=2000, knn=1, rr_seed=None):
+
+        self.alpha = alpha         # maximum number of selected features for each representative pointy
+        self.gamma = gamma         # impurity level (default: 0.2)
+        self.tau = tau             # number of iterations (default: 2)
+        self.sigma = sigma         # controls neighboring samples weighting (default: 1)
+        self.n_beta = n_beta       # number of distinct beta (default: 20)
+        self.nrrp = nrrp           # number of iterations for randomized rounding process (default: 2000)
+        self.knn = knn             # k nearest neighbours
+        self.rr_seed = rr_seed     # seed value for random rounding process
+
+    ###############################################################################################
+    # Fit (scikit learn pipeline function)
+    ### fits a model
+    # Input:
+    # -> training_data (X)   : The set of M by N features and observations
+    # -> training_labels (y) : The set of labels for each observation
+    # Output / Self assignment:
+    #    self.fstar : Selected features for each observation
+    ###############################################################################################
+
+    def fit(self, training_data, training_labels):
+
[…]
