### 2019-12-16 7183f82
Added manafest to allow the test data to be included and added an __ini__ file to the test directory
