+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include LFSpy\tests\matlab_Data.mat
