### 2019-12-16 073c729
update dependencies and fix a list
