### 2019-12-17 659b8b0
update test instructions
