+++ README.md
@@ -19,13 +19,13 @@ LFS requires:
 * [pytest](https://docs.pytest.org/en/latest/)>=5.0.0
 
 ### Testing
-We recommend running test.py after installing LFSpy to ensure the results obtained match expected outputs.
+We recommend running the provided test after installing LFSpy to ensure the results obtained match expected outputs.
 
 ```bash
-python test_lfs.py
+pytest --pyargs LFSpy
 ```
 
-This will output to console whether the results of LFSpy on two datasets (the sample dataset provided in this repository, and scikit-learn's Fisher Iris dataset) are exactly as expected, within 2%, or not even within 2%. 
+This will output to console whether or not the results of LFSpy on two datasets (the sample dataset provided in this repository, and scikit-learn's Fisher Iris dataset) are exactly as expected.
 
 So far, LFSpy has been tested on Windows 10 with and without Conda, and on Ubuntu. In all cases, results have been exactly the expected results.
 
