### 2019-12-19 a18abdd
Added sklearn parent classes
