### 2019-12-19 6953419
Added X and y validation in fit function. Transform training data within fit to match the algorithm. Scoring is now done by the inherited score function, test changed to reflect this.
