### 2020-01-17 McMasterRS/LFSpy#3 @effigies
[JOSS] Test failures
Installed as follows:

```
conda create -n lfspy python=3.7
conda activate lfspy
pip install lfspy
```

Tested using:

```
$ pytest --pyargs LFSpy
======================================= test session starts ========================================
platform darwin -- Python 3.7.6, pytest-5.3.3, py-1.8.1, pluggy-0.13.1
rootdir: /Users/markiewicz/Projects/joss-reviews/LFSpy
collected 2 items

tests/test_lfs.py FF                                                                         [100%]

============================================= FAILURES =============================================
_________________________________________ test_sample_data _________________________________________

    def test_sample_data():
        training_data, training_labels, testing_data, testing_labels = load_dataset('sample')
>       score, y_pred = train_model(training_data.T, training_labels, testing_data, testing_labels)

/System/Volumes/Data/anaconda3/envs/lfspy/lib/python3.7/site-packages/LFSpy/tests/test_lfs.py:97:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

x_train = array([[ 0.57089614, -0.5225384 , -0.44089233, ..., -1.56336955,
        -0.55169339,  1.48155663],
       [ 1.7703547...475, -0.58137331],
       [ 1.7703547 ,  3.9440912 ,  1.22648231, ..., -1.04889328,
        -1.1906316 , -0.27246217]])
y_train = array([0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0,
       0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1,...1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1,
       1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0], dtype=uint8)
x_test = array([[-0.62856242,  0.57089614, -0.62856242, ...,  0.57089614,
        -0.62856242,  0.57089614],
       [-0.5225384...694,  1.30829232],
       [-1.77183882,  0.71855258,  0.79467826, ..., -1.41023235,
        -1.58687822,  0.16628976]])
y_test = array([1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
       1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
      dtype=uint8)

    def train_model(x_train, y_train, x_test, y_test):
        '''
        Trains an tests and LFS model using default parameters on the given dataset.
        '''
        print('Training and testing an LFS model with default parameters.\nThis may take a few minutes...')
        lfs = LocalFeatureSelection(rr_seed=20)
        pipeline = Pipeline([('classifier', lfs)])
        pipeline.fit(x_train, y_train)
        y_pred = pipeline.predict(x_test)
        score = pipeline.score(x_test, y_test)

>       return score[0][0][0], y_pred
E       IndexError: invalid index to scalar variable.

/System/Volumes/Data/anaconda3/envs/lfspy/lib/python3.7/site-packages/LFSpy/tests/test_lfs.py:45: IndexError
--------------------------------------- Captured stdout call ---------------------------------------
Loading dataset sample...
Training and testing an LFS model with default parameters.
This may take a few minutes...
__________________________________________ test_iris_data __________________________________________

    def test_iris_data():
        training_data, training_labels, testing_data, testing_labels = load_dataset('iris')
>       score, y_pred = train_model(training_data, training_labels, testing_data.T, testing_labels)

/System/Volumes/Data/anaconda3/envs/lfspy/lib/python3.7/site-packages/LFSpy/tests/test_lfs.py:102:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

x_train = array([[5.1, 3.5, 1.4, 0.2],
       [4.7, 3.2, 1.3, 0.2],
       [5. , 3.6, 1.4, 0.2],
       [4.6, 3.4, 1.4, 0.3],
  ...],
       [5.8, 2.6, 4. , 1.2],
       [5.6, 2.7, 4.2, 1.3],
       [5.7, 2.9, 4.2, 1.3],
       [5.1, 2.5, 3. , 1.1]])
y_train = array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
       1, 1, 1, 1, 1, 1])
x_test = array([[4.9, 4.6, 5.4, 5. , 4.9, 4.8, 4.3, 5.7, 5.1, 5.1, 5.1, 5.1, 5. ,
        5.2, 4.7, 5.4, 5.5, 5. , 4.9, 5.1, 4....1. , 1.4, 1.5, 1.4, 1.4, 1. , 1.1, 1.3, 1.2, 1.4, 1.7,
        1. , 1. , 1.6, 1.6, 1.3, 1.3, 1.4, 1. , 1.2, 1.3, 1.3]])
y_test = array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
       1, 1, 1, 1, 1, 1])

    def train_model(x_train, y_train, x_test, y_test):
        '''
        Trains an tests and LFS model using default parameters on the given dataset.
        '''
        print('Training and testing an LFS model with default parameters.\nThis may take a few minutes...')
        lfs = LocalFeatureSelection(rr_seed=20)
        pipeline = Pipeline([('classifier', lfs)])
        pipeline.fit(x_train, y_train)
        y_pred = pipeline.predict(x_test)
        score = pipeline.score(x_test, y_test)

>       return score[0][0][0], y_pred
E       IndexError: invalid index to scalar variable.

/System/Volumes/Data/anaconda3/envs/lfspy/lib/python3.7/site-packages/LFSpy/tests/test_lfs.py:45: IndexError
--------------------------------------- Captured stdout call ---------------------------------------
Loading dataset iris...
Training and testing an LFS model with default parameters.
This may take a few minutes...
========================================= warnings summary =========================================
tests/test_lfs.py::test_sample_data
  /System/Volumes/Data/anaconda3/envs/lfspy/lib/python3.7/importlib/_bootstrap.py:219: RuntimeWarning: numpy.ufunc size changed, may indicate binary incompatibility. Expected 192 from C header, got 216 from PyObject
    return f(*args, **kwds)

-- Docs: https://docs.pytest.org/en/latest/warnings.html
============================= 2 failed, 1 warning in 616.09s (0:10:16) =============================
```

Environment: OSX/Conda
