### 2020-01-17 #1958 @effigies
## Review

LFSpy is a Python reimplementation of [LFS](https://github.com/armanfn/LFS), originally written by Narges Armanfard (@armanfn).

Much of the requested review is straightforward box-checking, so I'm restricting these comments to those where I have questions.

> ### General checks
>
> * [ ]  **Contribution and authorship:** Has the submitting author (@kiretd) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@kiretd wrote the paper and also contributed tests. According to the commit log, the contributors to the repository were:

```
    28  Oliver Cook <cooko@mcmaster.ca>
    23  Areeb Khawaja <khawaa1@mcmaster.ca>
    16  Kiret Dhindsa <kiretd@gmail.com>
    10  Tom Mudway <mudwayt@mcmaster.ca>
```

Areeb Khawaja appears to be missing from the author list, and Ron Harwood and Ranil Sonnadara have not contributed. Ranil appears to be a PI and Ron a staff member. I'll simply link to the [Authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) and would appreciate an affirmation that Ron and Ranil provided substantial contribution to the effort, and that @khawaa1 has elected not to be an author on this submission.

> ### Documentation

Documentation currently has the form of a [README.md](https://github.com/McMasterRS/LFSpy/blob/master/README.md). Although there are extensive comments in the code, there is not rendered documentation anywhere.

> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is currently missing

> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

This is there, although the installation requirements currently include `pytest`. The convention typically is to set installation requirements only to those needed to use the software, and to include tests and optional functionality as `extras_require` arguments to the `setuptools.setup` function. I will submit a pull request to the main repository to use this. (Authors are not obligated to accept this. Some projects prefer to treat tests as first-class requirements.)

> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The [example usage](https://github.com/McMasterRS/LFSpy#usage) is not directly runnable. From the perspective of users of [scikit-learn](https://scikit-learn.org), it is very likely enough, but fully runnable examples would be preferable, IMO. The [test_lfs.py](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/tests/test_lfs.py) module fetches two datasets. One of these could be used to flesh out the example usage.

> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

No. Examples of standard scikit-learn API documentation strings can be found [here](https://github.com/scikit-learn/scikit-learn/blob/b194674c4/sklearn/linear_model/_base.py#L375).

For these to be readable (as in [sklearn.linear_model.LinearRegression](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html#sklearn.linear_model.LinearRegression)), the documentation would need to be rendered.

> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

Yes, however, tests fail. Opened https://github.com/McMasterRS/LFSpy/issues/3.

> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This is missing.

> ### Software paper

> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

This currently feels implicit, and explicit comparisons to other methods would be helpful. I think comparisons to other classifiers within scikit-learn would serve equally as well as independent packages, if that's easier.

I suspect the authors are more familiar with the current state of the art in machine learning than I am, but if it's useful: My impression is that random forests (or other decision tree ensembles) will do a similar task, splitting the data into subspaces and selecting the most effective dimensions to distinguish classes.
