### 2020-01-17 a1fbc85
MNT: Update install_requires, package_data, version
