+++ setup.py
@@ -5,18 +5,24 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="LFSpy",
-    version="1.0.1",
+    version="1.0.2-dev",
     author="Oliver Cook, Kiret Dhindsa, Areeb Khawaja, Ron Harwood, Thomas Mudway",
-    install_requires=['numpy>=1.14', 'scipy>=1.1', 'scikit-learn>=0.18.2', 'pytest>=5.0.0'],
+    install_requires=['numpy>=1.14', 'scipy>=1.1', 'scikit-learn>=0.18.2'],
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/McMasterRS/LFS/",
     packages=setuptools.find_packages(),
-    include_package_data=True,
+    package_data={
+        'LFSpy': ['matlab_Data.mat'],
+        },
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: BSD License",
         "Operating System :: OS Independent",
     ],
+    tests_require=['pytest>=5.0.0'],
     python_requires='>=3.6',
-)   
+    extras_require={
+        'test': ['pytest>=5.0.0']
+        }
+)
