+++ MANIFEST.in
@@ -1,1 +1,0 @@
-include LFSpy\tests\matlab_Data.mat
