### 2020-01-21 23bbd16
Fixed test case - score is now correctly returned
