### 2020-01-21 cfeaa38
Transforming data in predict function
