### 2020-01-21 97d4e88
Adding comparison.py and figs
