### 2020-01-23 bbc1269
Create Readme.rst for Read the Docs Support
