+++ README.rst
@@ -1,0 +1,1 @@
+Initial readme file for read the docs (available here https://lfspy.readthedocs.io/en/latest/#)
