+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. LFSpy documentation master file, created by
+   sphinx-quickstart on Thu Jan 23 11:23:22 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to LFSpy's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
