+++ README.rst
@@ -1,1 +1,0 @@
-Initial readme file for read the docs (available here https://lfspy.readthedocs.io/en/latest/#)
