### 2020-01-23 37bebda
Added master_doc config to try to fix read the doc
