+++ docs/conf.py
@@ -38,6 +38,8 @@ templates_path = ['_templates']
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
+# Configure the master document for Read the Docs
+master_doc = 'index'
 
 # -- Options for HTML output -------------------------------------------------
 
