### 2020-01-23 69284e1
applied Read the Docs template sphinx_rtd_theme
