+++ docs/Functionality.rst
@@ -1,0 +1,2 @@
+Functionality
+=================================
