+++ docs/Examples.rst
@@ -1,0 +1,7 @@
+Examples
+=================================
+These are examples of how you can use LFSpy and the Localized feature selection method.
+
+Comparisons vs. other methods
+---------------------------------
+Potentially we could put some example of alternative methods here.
