+++ docs/index.rst
@@ -1,61 +1,78 @@
-.. LFSpy documentation master file, created by
-   sphinx-quickstart on Thu Jan 23 11:23:22 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to LFSpy's documentation!
+LFSpy
 =================================
 
 Localized feature selection (LFS) is a supervised machine learning approach for embedding localized feature selection in classification. The sample space is partitioned into overlapping regions, and subsets of features are selected that are optimal for classification within each local region. As the size and membership of the feature subsets can vary across regions, LFS is able to adapt to local variation across the entire sample space.
 
 This repository contains a python implementation of this method that is compatible with scikit-learn pipelines. For a Matlab version, refer to https://github.com/armanfn/LFS
 
+The LFS approach was developed by Nargus Armanfard. For further information please refer to:
+
+* N. Armanfard, JP. Reilly, and M. Komeili, "Local Feature Selection for Data Classification", IEEE Trans. on Pattern Analysis and Machine Intelligence, vol. 38, no. 6, pp. 1217-1227, 2016.
+* N. Armanfard, JP. Reilly, and M. Komeili, "Logistic Localized Modeling of the Sample Space for Feature Selection and Classification", IEEE Transactions on Neural Networks and Learning Systems, vol. 29, no. 5, pp. 1396-1413, 2018
+
+
 Installation
-=================================
-pip install lfspy
+---------------------------------
+
+LFSpy is available on the pypy distribution platform at https://pypi.org/project/LFSpy/.
+
+To install LFSpy along with its dependacies run the command:::
+
+    pip install lfspy
 
 Dependancies
-=================================
+---------------------------------
+
+
 LFS requires:
 
-Python 3
-NumPy>=1.14
-SciPy>=1.1
-Scikit-learn>=0.18.2
-pytest>=5.0.0
+* Python 3
+* NumPy>=1.14
+* SciPy>=1.1
+* Scikit-learn>=0.18.2
+* pytest>=5.0.0
 
 Testing
-=================================
+---------------------------------
+
+
 We recommend running the provided test after installing LFSpy to ensure the results obtained match expected outputs.
 
-pytest may be installed either directly through pip (pip install pytest) or using the test extra (pip install LFSpy[test]).
+pytest may be installed either directly through pip (pip install pytest) or using the test extra (pip install LFSpy[test]).::
 
-pytest --pyargs LFSpy
+    pytest --pyargs LFSpy
+
 This will output to console whether or not the results of LFSpy on two datasets (the sample dataset provided in this repository, and scikit-learn's Fisher Iris dataset) are exactly as expected.
 
 So far, LFSpy has been tested on Windows 10 with and without Conda, and on Ubuntu. In all cases, results have been exactly the expected results.
 
 Usage
-=================================
-To use LFSpy on its own:
+---------------------------------
 
-from LFSpy import LocalFeatureSelection
 
-lfs = LocalFeatureSelection()
-lfs.fit(training_data, training_labels)
-predicted_labels = lfs.predict(testing_data)
-total_error, class_error = lfs.score(testing_data, testing_labels)
-To use LFSpy as part of an sklearn pipeline:
+To use LFSpy on its own::
 
-from LFS import LocalFeatureSelection
-from sklearn.pipeline import Pipeline
+    from LFSpy import LocalFeatureSelection
 
-lfs = LocalFeatureSelection()
-pipeline = Pipeline([('lfs', lfs)])
-pipeline.fit(training_data, training_labels)
[…]
