+++ docs/Examples.rst
@@ -1,7 +1,3 @@
 Examples
 =================================
-These are examples of how you can use LFSpy and the Localized feature selection method.
-
-Comparisons vs. other methods
----------------------------------
-Potentially we could put some example of alternative methods here.
+TODO
