+++ docs/Community Guidelines.rst
@@ -1,2 +1,3 @@
 Community Guidelines
 =================================
+Contributions are welcomed and can be made to the public Git repository available at: https://github.com/McMasterRS/LFSpy
