+++ docs/Installation.rst
@@ -1,5 +1,19 @@
 Installation
 =================================
 
+LFSpy is available on the pypy distribution platform at https://pypi.org/project/LFSpy/.
+
+To install LFSpy along with its dependacies run the command:::
+
+    pip install lfspy
+
 Dependancies
 ---------------------------------
+
+LFS requires:
+
+* Python 3
+* NumPy>=1.14
+* SciPy>=1.1
+* Scikit-learn>=0.18.2
+* pytest>=5.0.0
