### 2020-01-29 f1ad616
Updates to the docs
