+++ paper.md
@@ -111,7 +111,7 @@ It can be seen that with the synthetic sample data, LFS outperforms the other tw
 
 ### Classification accuracies with the sample synthetic dataset
 
-![synth_data](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/SampleData_Results.png | width=256)
+![synth_data][results_sample]
 
 ### Classification accuracies with the Iris dataset
 
