### 2020-02-04 2ced938
fix md syntax in example
