+++ README.md
@@ -73,7 +73,7 @@ total_error, class_error = pipeline.score(testing_data, testing_labels)
 ### Example
 This example uses the sample data ([matlab_Data.mat](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/tests/matlab_Data.mat)) available in the [LFSpy/tests](https://github.com/McMasterRS/LFSpy/tree/master/LFSpy/tests) folder. The full example can be found in [example.py](https://github.com/McMasterRS/LFSpy/blob/master/example.py). On our test system, the fnial output prints the statement, "LFS test accuracy: 0.7962962962962963".
 
-'''
+```python
 import numpy as np
 from scipy.io import loadmat
 from LFSpy import LocalFeatureSelection
@@ -92,7 +92,7 @@ pipeline.fit(x_train, y_train)
 y_pred = pipeline.predict(x_test)
 score = pipeline.score(x_test, y_test)
 print('LFS test accuracy: {}'.format(score))
-'''
+```
 
 ## Authors
 *  Oliver Cook
