### 2020-02-05 17487ab
adding contribution guidelines and docs hyperlink
