### 2020-02-06 b05acf8
Updating comparisons and IrisData_results
