### 2020-02-06 790a479
update comparison section
