### 2020-02-06 874d548
small comparisons update
