### 2020-02-06 McMasterRS/LFSpy#3 @cooko-mcmaster
It appears that inside the unit test function 'test_iris_data' the score was incorrectly being indexed when returned. This may have been accidentally carried over from a previous test version. The score is now simply a single value rather than any sort of collection.

I removed the indexing and the tests now pass as expected.
