### 2020-02-06 e3f3846
testing html syntax for image resizing
