### 2020-02-06 34de710
checking another method + controlling aspect ratio
