+++ paper.md
@@ -112,12 +112,12 @@ It can be seen that with both datasets LFS outperforms the other two methods, pa
 ### Classification accuracies with the sample synthetic dataset
 
 <!-- ![synth_data][results_sample] -->
-![synth_data](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/SampleData_Results.png)
+![synth_data](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/SampleData_Results.png) <!-- .element height="75%" width="75%" -->
 
 ### Classification accuracies with the Iris dataset
 
 <!-- ![iris_data][results_iris] -->
-<img src="https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/IrisData_Results.png" width="512" height="512">
+<img src="https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/IrisData_Results.png" width="75%" height="75%">
 
 # Acknowledgments
 Funding for this project was obtained through the CANARIE Research Software Program Local Support Initiative.
