### 2020-02-06 6a8d1bd
comparing md method to html method
