+++ paper.md
@@ -112,7 +112,7 @@ It can be seen that with both datasets LFS outperforms the other two methods, pa
 ### Classification accuracies with the sample synthetic dataset
 
 <!-- ![synth_data][results_sample] -->
-![synth_data](https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/SampleData_Results.png){:height="75%" width="75%"}
+<img src="https://github.com/McMasterRS/LFSpy/blob/master/LFSpy/comparisons/SampleData_Results.png" width="75%" height="75%">
 
 ### Classification accuracies with the Iris dataset
 
