### 2020-02-06 #1958 @kiretd
@dfm Hey, sorry about this, but I think I need help including figures in the pdf. Using standard markdown syntax (same as the example given on the submission instructions page), whedon spits out the error a few posts above. I switched to html because it was also the only method that allowed me to resize the images, and now the pdf compiles without images. Hopefully I'm just missing something really obvious?
