+++ paper/paper.md
@@ -115,11 +115,11 @@ It can be seen that with both datasets LFS outperforms the other two methods, pa
 
 ### Classification accuracies with the sample synthetic dataset
 
-![Synthetic dataset](results_sample)
+![Synthetic dataset](SampleData_Results.png)
 
 ### Classification accuracies with the Iris dataset
 
-![Iris data](results_iris)
+![Iris data](IrisData_Results.png)
 
 # Acknowledgments
 Funding for this project was obtained through the CANARIE Research Software Program Local Support Initiative.
