### 2020-02-06 #1958 @kiretd
@dfm Thanks so much. It looks like the images just needed to be in the same folder for whedon to find them?
Also, is there no way to resize the images with markdown syntax? I assume the latex conversion won't work if the images are included using html syntax.
