### 2020-02-07 43ae6a3
add missing period in second figure caption
