### 2020-02-10 #1958 @kiretd
Just small update to add the figure captions
