### 2020-02-13 #1958 @sauln
I'll have to provide a more thorough review near the beginning of next week, but I can supply a few quick notes from pursuing the documentation:
- The docs link in the readme doesn't work. You'll need to use `[lfspy.readthedocs.io](https://lfspy.readthedocs.io)`
- It looks like you manually wrote the Functionality page and I see a lot of good comments in the code. If you reformatted the docstrings to comply with an [autodocs](http://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) format, they would render with the `__doc__` magic method, and you could use autodocs to render the functionality page.
- It looks like each section in the [homepage](https://lfspy.readthedocs.io/en/latest/index.html) of the docs is repeated as an individual page.  Is this necessary?
- The contribution guide says that contributions should follow PEP8, but the codebase looks like it doesn't follow PEP8 itself.  Have you ran a linter or autoformatter over code? I highly recommend [black](https://black.readthedocs.io/en/stable/) but any will do.
- Could you provide more examples with larger data sets?
- The table of tunable parameters in the docs homepage hasn't rendered correctly.
- The tests appear to be end-to-end tests.  I would like to see more unit tests that focus on individual pieces of logic.
