### 2020-02-26 #1958 @dfm
Hi team! Just checking in here on the status.

@kiretd: I wanted to double check that you got @sauln's recommendations. Let us know if you have any questions!

@effigies: It seems like there might be some remaining issues for you. Can you let us know if there's more that you would like to see beyond this response: https://github.com/openjournals/joss-reviews/issues/1958#issuecomment-583011174

Thanks!
