### 2020-02-28 6c7a4c9
Loosened style guide in docs
