### 2020-02-28 0ff3233
DOC: Typos, style and small suggestions
