+++ docs/Functionality.rst
@@ -1,5 +1,5 @@
 Functionality
-=================================
+=============
 
 ::
 
@@ -46,7 +46,7 @@ Functionality
 +----------------+-----------------------------------------------------------------------------+
 
 Methods
----------------------------------
+-------
 
 +-----------------------------------------------------+-----------------------+
 | fit(self, training_data, training_labels)           |                       |
