### 2020-02-28 #1958 @effigies
@dfm My concerns have been addressed.

I submitted a minor PR (https://github.com/McMasterRS/LFSpy/pull/6) for some cleanups in the documentation, but that should not be considered blocking.
