### 2020-02-29 #1958 @kiretd
Hi @dfm
Thanks for checking in. Yes, I did get the recommendations by @sauln (thanks for those @sauln). I just moved continents last week, so I haven’t gotten much done on this myself yet. The team has been working on some of the recommendations in the meantime, and hopefully we’ll have an update soon.

@sauln:
Thanks for your suggestions! I’ll respond one by one.

> * The docs link in the readme doesn't work. You'll need to use `[lfspy.readthedocs.io](https://lfspy.readthedocs.io)`

Thanks, looks fixed on our end now.

> * It looks like you manually wrote the Functionality page and I see a lot of good comments in the code. If you reformatted the docstrings to comply with an [autodocs](http://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) format, they would render with the `__doc__` magic method, and you could use autodocs to render the functionality page.

This sounds really great, and I think we’ll likely start implementing it for our projects going forward. I’m not sure there’s much value in spending the time to redo our comments and functionality page for LFSpy right now, but we’re taking a look at whether it makes sense to switch it over gradually over the next few months or so.

> * It looks like each section in the [homepage](https://lfspy.readthedocs.io/en/latest/index.html) of the docs is repeated as an individual page.  Is this necessary?

Maybe not necessary, but the way we have our hyperlinked table of contents set up is that each section has a link on the sidebar for easier navigation.

> * The contribution guide says that contributions should follow PEP8, but the codebase looks like it doesn't follow PEP8 itself.  Have you ran a linter or autoformatter over code? I highly recommend [black](https://black.readthedocs.io/en/stable/) but any will do.

Oliver from our team ran it through Black as you recommended. Pretty cool tool! I’m not an expert on this part, but I believe that it’s not exactly formatted correctly even after using Black. For now, we’ve just removed the PEP8 guideline so there’s no inconsistency.

> * Could you provide more examples with larger data sets?

Do you have any particular datasets in mind? Typically sample datasets used for illustration are small so that the tests and examples can be run relatively quickly. In addition, LFS is particularly well suited for p>>n and small n problems, as per our statement of need, so we think our examples (one custom, the other the standard sk-learn sample dataset used in many of their examples) are rather suitable.

> * The table of tunable parameters in the docs homepage hasn't rendered correctly.

This looks to be fixed on our end. Can you confirm it looks right on your end too?

> * The tests appear to be end-to-end tests.  I would like to see more unit tests that focus on individual pieces of logic.

LFSpy is meant to be run as a single function, usually in a pipeline, but at least as a classifier or a feature selector (using the same function). Do you mean we should publish tests of internal functions as would be used for internal validation?
