### 2020-03-11 #1958 @kiretd
> I submitted a minor PR ([McMasterRS/LFSpy#6](https://github.com/McMasterRS/LFSpy/pull/6)) for some cleanups in the documentation, but that should not be considered blocking.

Thanks @effigies looks good to me, just waiting for one more person to review in case I missed something.

@dfm Just want to check in to see where we are with the process. I know I took some time to respond while we made the updates, so hopefully nothing is waiting on me right now.
