### 2020-03-11 #1958 @dfm
@sauln: Do you feel that @kiretd has addressed some of your concerns? Do you have more feedback?

@kiretd: In response to your question to @sauln:

> Do you mean we should publish tests of internal functions as would be used for internal validation?

I believe that that was what he was saying. In general, [unit tests](https://en.wikipedia.org/wiki/Unit_testing) should test each sub-component of the code, not just the end-to-end result. Unit testing, specifically, is not a *requirement* for JOSS, but it is recommended best practice.
