### 2020-03-11 #1958 @kiretd
@dfm and @sauln Thanks, yep, unit testing makes sense. I'm going to look into whether we have the resources to add that in now, or if we'll need to add it in gradually. I'm looking into exactly what would be the best way to do it now.
