### 2020-03-27 c21e5fb
proper docstrings, non-inlined comments, and removed unused variables
