+++ LFSpy/LFS.py
@@ -69,15 +69,7 @@ class LocalFeatureSelection(ClassifierMixin, BaseEstimator):
     """
 
     def __init__(
-        self,
-        alpha=19,
-        gamma=0.2,
-        tau=2,
-        sigma=1,
-        n_beta=20,
-        nrrp=2000,
-        knn=1,
-        rr_seed=None,
+        self, alpha=19, gamma=0.2, tau=2, sigma=1, n_beta=20, nrrp=2000, knn=1, rr_seed=None,
     ):
         self.alpha = alpha
         self.gamma = gamma
@@ -106,14 +98,9 @@ class LocalFeatureSelection(ClassifierMixin, BaseEstimator):
         self.fstar = np.zeros(
             training_data.shape
         )  # selected features for each representative point; if fstar(i, j) = 1, ith feature is selected for jth representative point.
-        self.fstar_lin = np.zeros(
-            training_data.shape
-        )  # fstar before applying randomized rounding process
+        self.fstar_lin = np.zeros(training_data.shape)  # fstar before applying randomized rounding process
 
-        (
-            m_features,
-            n_observations,
-        ) = training_data.shape  # (M, N) M number of candidate features, N observations
+        (m_features, n_observations,) = training_data.shape  # (M, N) M number of candidate features, N observations
         n_total_cls = [
             np.sum(training_labels ^ 1),
             np.sum(training_labels),
@@ -153,81 +140,47 @@ class LocalFeatureSelection(ClassifierMixin, BaseEstimator):
                 for i in range(n_observations):
                     # total distance along features selected by fstar
                     fstar_feature_dist_0 = np.sqrt(
-                        np.sum(
-                            observation_distances[
-                                :, excluding_selected & (training_labels == 0)
-                            ]
-                            * fstar_mask[:, i][..., None],
-                            axis=0,
-                        )
+                        np.sum(observation_distances[:, excluding_selected & (training_labels == 0)] * fstar_mask[:, i][..., None], axis=0,)
                     )
                     # total distance along features selected by fstar
                     fstar_feature_dist_1 = np.sqrt(
-                        np.sum(
-                            observation_distances[
-                                :, excluding_selected & (training_labels == 1)
-                            ]
-                            * fstar_mask[:, i][..., None],
-                            axis=0,
-                        )
+                        np.sum(observation_distances[:, excluding_selected & (training_labels == 1)] * fstar_mask[:, i][..., None], axis=0,)
                     )
-                    w11 = np.exp(
-                        (-((fstar_feature_dist_1 - np.min(fstar_feature_dist_1)) ** 2))
-                        / self.sigma
-                    )
-                    w22 = np.exp(
-                        (-((fstar_feature_dist_0 - np.min(fstar_feature_dist_0)) ** 2))
-                        / self.sigma
-                    )
-                    observation_weight[
-                        i, excluding_selected & (training_labels == 0)
-                    ] = w22
-                    observation_weight[
-                        i, excluding_selected & (training_labels == 1)
-                    ] = w11
+                    w11 = np.exp((-((fstar_feature_dist_1 - np.min(fstar_feature_dist_1)) ** 2)) / self.sigma)
+                    w22 = np.exp((-((fstar_feature_dist_0 - np.min(fstar_feature_dist_0)) ** 2)) / self.sigma)
+                    observation_weight[i, excluding_selected & (training_labels == 0)] = w22
+                    observation_weight[i, excluding_selected & (training_labels == 1)] = w11
 
                 # mean weight of all observations or features
                 average_observation_weight = np.mean(observation_weight, axis=0)
 
                 # normalized weight for all observations not including the current observation
                 normalized_weight = np.zeros((1, n_observations))
-                normalized_weight[:, training_labels == 0] = average_observation_weight[
-                    training_labels == 0
[…]
