### 2020-03-27 #1958 @sauln
Sorry for the delay. I was on PTO, and then the whole global pandemic thing.

Many of my suggested edits were taken care of. The most important of those that were not addressed, I addressed myself in a PR:
[#7](https://github.com/McMasterRS/LFSpy/pull/7) establishes proper docstrings so the code is self-documenting, and
[#8](https://github.com/McMasterRS/LFSpy/pull/8) tweaks the `black` parameters so the formatting isn't so strange looking.

I highly suggest breaking the `fit` method into smaller pieces. At its deepest, I count 14 levels of indentation.  That suggests to me the library will be very difficult to maintain moving forward.

> In addition, LFS is particularly well suited for p>>n and small n problems, as per our statement of need, so we think our examples (one custom, the other the standard sk-learn sample dataset used in many of their examples) are rather suitable.

The iris data set has p=4 and n=150. Maybe the scikit-learn wine dataset would be a richer example with 13 dimensions? Another option would be to add a handful of examples using the [scikit-learn generated datasets API](https://scikit-learn.org/stable/datasets/index.html#generated-datasets). Then you could really show off the strengths.
