### 2020-03-27 #1958 @sauln
The above are just strong suggestions @dfm.  If my two PRs are merged, I believe my concerns would be sufficiently addressed.

Thank you.
