### 2020-04-14 #1958 @kiretd
Hi @dfm and @sauln, thanks for the updates!

We did indeed merge the changes Saul generously provided, and my delay in response is because we're still working on some of the others (though they were listed as optional). I had a draft response, so I'll just put it below so it's clear where we are now.

> Many of my suggested edits were taken care of. The most important of those that were not addressed, I addressed myself in a PR:
> [#7](https://github.com/McMasterRS/LFSpy/pull/7) establishes proper docstrings so the code is self-documenting, and
> [#8](https://github.com/McMasterRS/LFSpy/pull/8) tweaks the `black` parameters so the formatting isn't so strange looking.

If I remember correctly, I think we tried to do this, but it didn't look quite right. It looks like you fixed the issue with it, so thanks!

> I highly suggest breaking the `fit` method into smaller pieces. At its deepest, I count 14 levels of indentation. That suggests to me the library will be very difficult to maintain moving forward.

This is actually something we've been working on, and updates on this will come. The way it is now is definitely far from best coding practice, and it won't stay that way. It will take us some time though.

> > In addition, LFS is particularly well suited for p>>n and small n problems, as per our statement of need, so we think our examples (one custom, the other the standard sk-learn sample dataset used in many of their examples) are rather suitable.
>
> The iris data set has p=4 and n=150. Maybe the scikit-learn wine dataset would be a richer example with 13 dimensions? Another option would be to add a handful of examples using the [scikit-learn generated datasets API](https://scikit-learn.org/stable/datasets/index.html#generated-datasets). Then you could really show off the strengths.

That is correct, and all four features are discriminative. We add increasing numbers of non-discriminative features to showcase exactly what LFSpy excels at (p>>n, but most of those p are not actually useful) so we can simulate a noise dataset. I'm now also building a similar example using scikit-learn's sklearn.datasets.make_classification function, which does have some nice options for controlling the properties of generated datasets. It's taking a bit of time, but I'm hoping it'll be added to the paper and the docs by the end of the week.

In the meantime, it would be great to know if there's anything else we should work on prior to publication. Thanks!
