### 2020-04-17 37da95e
adding text for new experiment
just the text for now, new figure soon to come.
