### 2020-04-20 #1958 @kiretd
@dfm @sauln
I've updated the paper and the repo with a new experiment that I think will satisfy one of the requests.
Cheers all, hope you're in good health.
