### 2020-05-04 #1958 @dfm
@kiretd: Thanks for your patience! I've just submitted two pull requests with small changes to the paper. Once you've accepted those, take a last proof read, then:

1. Bump the version number and tag a new release of the code
2. Create an archive of the code using Zenodo or something similar
3. Make sure that the metadata (title and author list, for example) of the archive match the JOSS paper (on Zenodo you can edit the metadata after creating the archive)

Once you've completed those steps, please post the new version number and the DOI for the archive to this thread. Thanks!
