+++ paper/paper.bib
@@ -42,3 +42,15 @@
   publisher={IEEE},
   doi={10.1109/JBHI.2018.2877738}
 }
+
+@article{Boshra:2019,
+  title={From group-level statistics to single-subject prediction: machine learning detection of concussion in retired athletes},
+  author={Boshra, Rober and Dhindsa, Kiret and Boursalie, Omar and Ruiter, Kyle I and Sonnadara, Ranil and Samavi, Reza and Doyle, Thomas E and Reilly, James P and Connolly, John F},
+  journal={IEEE Transactions on Neural Systems and Rehabilitation Engineering},
+  volume={27},
+  number={7},
+  pages={1492--1501},
+  year={2019},
+  publisher={IEEE},
+  doi={10.1109/TNSRE.2019.2922553}
+}
