### 2020-05-07 #1958 @kiretd
Thanks @dfm, as well as @effigies and @sauln for all of the help and effort!

1. We just tagged a release of the code. It's at version 1.0 on Github, but version 1.0.4 on Pypi. Do these need to be the same?
2. DOI: 10.5281/zenodo.3813708
3. Looks right to me.

I've made all of the changes I want to the paper. Did I provide you with everything you need?
