### 2020-05-07 #1958 @dfm
It's good practice to synchronize versions across PyPI and GitHub releases, but it's not absolutely necessary.

That DOI isn't resolving and I can't find the archive on Zenodo manually. Can you share the Zenodo link directly?
