### 2020-05-07 #1958 @effigies
DOI resolves for me. I've noticed Zenodo DOIs have a bit of an increased lag before they resolve, lately.
