### 2020-05-07 #1958 @dfm
Great! Yes - it's resolving now. Please update the metadata in the Zenodo archive so that the title matches the paper: "LFSpy: A Python Implementation of Local Feature Selection for Data Classification with scikit-learn Compatibility".
