### 2020-05-10 #1958 @kiretd
@dfm, @effigies, @sauln
Thanks all! Now that it's all done, I just want to say that this is easily the most transparent, innovative, and overall best review phase I've been through with a journal. I really hope this is what peer review evolves into as the years go on. Not just that, but you all were also extremely helpful and educational throughout the process, so really well done.
